+++ addOns/saverawmessage/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Retired
+- This add-on has been retired, and its functionality has been replaced by the Import/Export Add-on.
 
 ## [6] - 2021-10-06
 ### Changed