+++ addOns/savexmlmessage/CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-
+### Retired
+- This add-on has been retired, and its functionality has been replaced by the Import/Export Add-on.
 
 ## [0.2.0] - 2021-10-06
 ### Changed